+++ reuseport_other.go
@@ -1,7 +1,0 @@
-//go:build !linux
-
-package main
-
-func setSOReuseport(fd int) error {
-	return nil
-}